将菜单相关的类转移到单独的包里方便管理，以便于某些特殊需求
